simplify eventual hook
