+++ packages/@eventual/core-runtime/src/eventual-hook.ts
@@ -1,5 +1,17 @@
-import { type EventualHook } from "@eventual/core/internal";
+import {
+  EventualPromise,
+  PropertyType,
+  type Call,
+  type CallOutput,
+  type EventualHook,
+  type Property,
+} from "@eventual/core/internal";
 import { AsyncLocalStorage } from "async_hooks";
+import type { CallExecutor } from "./call-executor.js";
+import {
+  PropertyRetriever,
+  getEventualProperty,
+} from "./property-retriever.js";
 
 declare global {
   // eslint-disable-next-line no-var
@@ -25,9 +37,35 @@ if (!globalThis.eventualCallHookStore) {
   globalThis.eventualCallHookStore = new AsyncLocalStorage<EventualHook>();
 }
 
+export class DefaultEventualHook implements EventualHook {
+  constructor(
+    private executor: CallExecutor,
+    private propertyRetriever: PropertyRetriever
+  ) {}
+
+  public executeEventualCall<P extends Call>(
+    eventual: P
+  ): EventualPromise<any> {
+    return this.executor.execute(eventual) as CallOutput<P>;
+  }
+
+  public getEventualProperty<P extends Property = Property>(
+    property: P
+  ): PropertyType<P> {
+    return getEventualProperty(
+      property,
+      this.propertyRetriever
+    ) as PropertyType<P>;
+  }
+}
+
 export async function enterEventualCallHookScope<R>(
-  eventualHook: EventualHook,
+  callExecutor: CallExecutor,
+  propertyRetriever: PropertyRetriever,
   callback: () => R
 ): Promise<Awaited<R>> {
-  return await globalThis.eventualCallHookStore.run(eventualHook, callback);
+  return await globalThis.eventualCallHookStore.run(
+    new DefaultEventualHook(callExecutor, propertyRetriever),
+    callback
+  );
 }
